Restructure operator definitions and syntax support
This closes #20.
