+++ core/src/main/scala/com/danielwestheide/kontextfrei/DCollectionConstructors.scala
@@ -1,0 +1,8 @@
+package com.danielwestheide.kontextfrei
+
+import scala.collection.immutable.Seq
+import scala.reflect.ClassTag
+
+private[kontextfrei] trait DCollectionConstructors[DCollection[_]] {
+  def unit[A: ClassTag](as: Seq[A]): DCollection[A]
+}
