+++ core/src/main/scala/com/danielwestheide/kontextfrei/rdd/RDDBase.scala
@@ -1,0 +1,7 @@
+package com.danielwestheide.kontextfrei.rdd
+
+import org.apache.spark.SparkContext
+
+private[kontextfrei] trait RDDBase {
+  def sparkContext: SparkContext
+}
