+++ core/src/main/scala/com/danielwestheide/kontextfrei/stream/StreamOps.scala
@@ -1,0 +1,9 @@
+package com.danielwestheide.kontextfrei.stream
+
+import com.danielwestheide.kontextfrei.DCollectionOps
+
+class StreamOps
+    extends DCollectionOps[Stream]
+    with StreamBaseFunctions
+    with StreamPairFunctions
+    with StreamConstructors
