+++ core/src/main/scala/com/danielwestheide/kontextfrei/stream/StreamOpsSupport.scala
@@ -1,0 +1,10 @@
+package com.danielwestheide.kontextfrei.stream
+
+import com.danielwestheide.kontextfrei.DCollectionOps
+
+trait StreamOpsSupport {
+  implicit val streamCollectionOps: DCollectionOps[Stream] =
+    new StreamOps
+}
+
+object StreamOpsSupport extends StreamOpsSupport
