+++ core/src/main/scala/com/danielwestheide/kontextfrei/syntax/Imports.scala
@@ -1,0 +1,8 @@
+package com.danielwestheide.kontextfrei.syntax
+
+trait Imports
+    extends DistributionOps
+    with SyntaxSupport
+    with DistributionSyntaxSupport
+
+object Imports extends Imports
