+++ core/src/test/scala/com/danielwestheide/kontextfrei/RDDCollectionOpsSpec.scala
@@ -1,5 +1,6 @@
 package com.danielwestheide.kontextfrei
 
+import com.danielwestheide.kontextfrei.rdd.RDDOpsSupport
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.scalatest.BeforeAndAfterAll
@@ -9,7 +10,7 @@ class RDDCollectionOpsSpec
     with BeforeAndAfterAll {
   implicit val sparkContext = new SparkContext("local[2]", "dcollection-spec")
   override implicit val ops: DCollectionOps[RDD] =
-    RDDCollectionOps.rddCollectionOps
+    RDDOpsSupport.rddCollectionOps
   override protected def afterAll(): Unit = {
     sparkContext.stop()
   }
