+++ core/src/test/scala/com/danielwestheide/kontextfrei/StreamCollectionOpsSpec.scala
@@ -1,6 +1,8 @@
 package com.danielwestheide.kontextfrei
 
+import com.danielwestheide.kontextfrei.stream.StreamOpsSupport
+
 class StreamCollectionOpsSpec extends DCollectionOpsProperties[Stream] {
   override implicit val ops: DCollectionOps[Stream] =
-    StreamCollectionOps.streamCollectionOps
+    StreamOpsSupport.streamCollectionOps
 }
