+++ scalatest/src/main/scala/com/danielwestheide/kontextfrei/scalatest/RDDSpec.scala
@@ -1,6 +1,7 @@
 package com.danielwestheide.kontextfrei.scalatest
 
-import com.danielwestheide.kontextfrei.{DCollectionOps, RDDCollectionOps}
+import com.danielwestheide.kontextfrei.DCollectionOps
+import com.danielwestheide.kontextfrei.rdd.RDDOpsSupport
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.scalatest.{BeforeAndAfterAll, Suite}
@@ -9,7 +10,7 @@ trait RDDSpec extends KontextfreiSpec[RDD] with BeforeAndAfterAll {
   self: Suite =>
   implicit val sparkContext = new SparkContext("local[2]", "my-test")
   override implicit val ops: DCollectionOps[RDD] =
-    RDDCollectionOps.rddCollectionOps
+    RDDOpsSupport.rddCollectionOps
   override protected def afterAll(): Unit = {
     sparkContext.stop()
   }
