+++ scalatest/src/main/scala/com/danielwestheide/kontextfrei/scalatest/StreamSpec.scala
@@ -1,7 +1,7 @@
 package com.danielwestheide.kontextfrei.scalatest
 
-import com.danielwestheide.kontextfrei.StreamCollectionOps
+import com.danielwestheide.kontextfrei.stream.StreamOpsSupport
 
 trait StreamSpec extends KontextfreiSpec[Stream] {
-  override implicit def ops = StreamCollectionOps.streamCollectionOps
+  override implicit def ops = StreamOpsSupport.streamCollectionOps
 }
